requirements.txt
# Core ML stack
torch==2.2.1
torchvision==0.17.1
ultralytics==8.1.28

# Vision + image processing
opencv-python==4.9.0.80
pillow==10.3.0
numpy==1.26.4

# Classifier + safetensors
safetensors==0.4.2

# Metadata + XMP
pyexiftool==0.5.6

# Downloader
requests==2.31.0

# Utilities
tqdm==4.66.2
rich==13.7.1
huggingface_hub==0.23.4

# Testing
pytest==9.0.3
pytest-order==1.4.0




